# ~1.5 times faster triton kernel