This PR improves the Triton kernel from ~12 seconds in #1 to ~8.5 seconds on 6,553,600 samples with 65,536 clusters with near identical NMI.

FastKMeans Trition backend is now 4-5 times faster than FAISS on a single GPU.

4090 Benchmarks:
<details>
  <summary>6,553,600 samples with 65,536 clusters</summary>

```
=== Faiss KMeans on 6553600 samples, 65536 clusters ===
[Faiss KMeans] Done in 48.9896 seconds
[Faiss KMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9792

=== FastKMeans on 6553600 samples, 65536 clusters ===
[PyTorch FastKMeans] Done in 40.1800 seconds
[PyTorch FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9791

=== FastKMeans on 6553600 samples, 65536 clusters ===
[Triton FastKMeans] Done in 8.4807 seconds
[Triton FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9790
```

</details>

<details>
  <summary>13,107,200 samples with 131,072 clusters</summary>

```
=== Faiss KMeans on 13107200 samples, 131072 clusters ===
[Faiss KMeans] Done in 189.3853 seconds
[Faiss KMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9801

=== FastKMeans on 13107200 samples, 131072 clusters ===
[PyTorch FastKMeans] Done in 156.6707 seconds
[PyTorch FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9803

=== FastKMeans on 13107200 samples, 131072 clusters ===
[Triton FastKMeans] Done in 30.8149 seconds
[Triton FastKMeans] Evaluation Metrics:
  Normalized Mutual Info (NMI): 0.9802
```

</details>